Support for BATCH Certific@2 and fix some BUGs with CONTRACT  Cetrific@2 . I've gone to the dark side, ALTER, Hibernate, JSFs :-(

+++ aon-ui-sepe/src/main/java/com/esferalia/aon/ui/sepe/file/CertificadosWriter.java
@@ -370,7 +370,7 @@ public class CertificadosWriter implements Serializable {
 		
 		if ( batchDetail.getSuspensionCause() == SuspensionCause.C18 ) {
 			try {
-				Double coeficenteReduccion = Double.parseDouble(ereFactor.getExpression()) * 100.00;
+				Integer coeficenteReduccion = (int)Double.parseDouble(ereFactor.getExpression()) * 100;
 				o.setPorcentualReduccionERE(completeLength(coeficenteReduccion.toString(),4,false));
 			} catch ( Exception e ) {
 				o.setPorcentualReduccionERE(null);
